+++ equinox-web-csharp/Web/CosmosContext.cs
@@ -64,7 +64,7 @@ namespace TodoBackendTemplate
                 ? null
                 : CachingStrategy.NewSlidingWindow(_cache, TimeSpan.FromMinutes(20));
             var cat = new CosmosStoreCategory<TEvent, TState, Unit>(_store, codec.ToJsonElementCodec(), FuncConvert.FromFunc(fold), initial, cacheStrategy, accessStrategy, compressUnfolds:FSharpOption<bool>.None);
-            return cat.Resolve(handlerLog);
+            return args => cat.Resolve(handlerLog).Invoke(args.Item1, args.Item2);
         }
     }
 }