+++ feed-source/Domain/TicketsEpoch.fs
@@ -6,7 +6,7 @@
 module FeedSourceTemplate.Domain.TicketsEpoch
 
 let [<Literal>] Category = "TicketsEpoch"
-let streamName (fcId, epochId) = struct (Category, FsCodec.StreamName.createStreamId [FcId.toString fcId; TicketsEpochId.toString epochId])
+let streamId = Equinox.StreamId.gen2 FcId.toString TicketsEpochId.toString
 
 // NB - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 [<RequireQualifiedAccess>]
@@ -68,7 +68,7 @@ let decide capacity candidates (currentIds, closed as state) =
         { accepted = addedItemIds; residual = residualItems; content = currentIds; closed = closed }, events
 
 /// Service used for the write side; manages ingestion of items into the series of epochs
-type IngestionService internal (capacity, resolve : FcId * TicketsEpochId -> Equinox.Decider<Events.Event, Fold.State>) =
+type IngestionService internal (capacity, resolve : struct (FcId * TicketsEpochId) -> Equinox.Decider<Events.Event, Fold.State>) =
 
     /// Handles idempotent deduplicated insertion into the set of items held within the epoch
     member _.Ingest(fcId, epochId, ticketIds) : Async<Result> =
@@ -86,11 +86,11 @@ type IngestionService internal (capacity, resolve : FcId * TicketsEpochId -> Equ
 module Config =
 
     let private create_ capacity resolve =
-        IngestionService(capacity, streamName >> resolve)
+        IngestionService(capacity, streamId >> resolve)
     let private (|Category|) = function
         | Config.Store.Memory store ->            Config.Memory.create Events.codec Fold.initial Fold.fold store
         | Config.Store.Cosmos (context, cache) -> Config.Cosmos.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
-    let create capacity (Category cat) = Config.createDecider cat |> create_ capacity
+    let create capacity (Category cat) = Config.createDecider cat Category |> create_ capacity
 
 /// Custom Fold and caching logic compared to the IngesterService
 /// - When reading, we want the full Items
@@ -108,7 +108,7 @@ module Reader =
 
     type StateDto = { closed : bool; tickets : Events.Item[] }
 
-    type Service internal (resolve : FcId * TicketsEpochId -> Equinox.Decider<Events.Event, State>) =
+    type Service internal (resolve : struct (FcId * TicketsEpochId) -> Equinox.Decider<Events.Event, State>) =
 
         /// Returns all the items currently held in the stream
         member _.Read(fcId, epochId) : Async<StateDto> =
@@ -120,4 +120,4 @@ module Reader =
         let private (|Category|) = function
             | Config.Store.Memory store ->            Config.Memory.create Events.codec initial fold store
             | Config.Store.Cosmos (context, cache) -> Config.Cosmos.createUnoptimized Events.codec initial fold (context, cache)
-        let create (Category cat) = Service(streamName >> Config.createDecider cat)
+        let create (Category cat) = Service(streamId >> Config.createDecider cat Category)