+++ propulsion-cosmos-reactor/Todo.fs
@@ -3,7 +3,7 @@ module ReactorTemplate.Todo
 open Propulsion.Internal
 
 let [<Literal>] Category = "Todos"
-let streamName (clientId : ClientId) = struct (Category, ClientId.toString clientId)
+let streamId = Equinox.StreamId.gen ClientId.toString
 let [<return: Struct>] (|StreamName|_|) = function FsCodec.StreamName.CategoryAndId (Category, ClientId.Parse clientId) -> ValueSome clientId | _ -> ValueNone
 
 // NB - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
@@ -67,5 +67,5 @@ module Config =
 
     let private (|Category|) = function
         | Config.Store.Cosmos (context, cache) -> Config.Cosmos.createSnapshotted Events.codecJe Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
-    let create (Category cat) = streamName >> Config.createDecider cat |> Service
+    let create (Category cat) = Service(streamId >> Config.createDecider cat Category)