+++ propulsion-tracking-consumer/SkuSummary.fs
@@ -1,7 +1,7 @@
 ﻿module ConsumerTemplate.SkuSummary
 
 let [<Literal>] Category = "SkuSummary"
-let streamName (id : SkuId) = struct (Category, SkuId.toString id)
+let streamId = Equinox.StreamId.gen SkuId.toString
 
 // NB - these types and the union case names reflect the actual storage formats and hence need to be versioned with care
 module Events =
@@ -63,4 +63,4 @@ module Config =
 
     let private (|Category|) = function
         | Config.Store.Cosmos (context, cache) -> Config.Cosmos.createSnapshotted Events.codec Fold.initial Fold.fold (Fold.isOrigin, Fold.toSnapshot) (context, cache)
-    let create (Category cat) = Service(streamName >> Config.createDecider cat)
+    let create (Category cat) = Service(streamId >> Config.createDecider cat Category)